+++ sourcespec/source_spec.py
@@ -81,6 +81,8 @@ def main():
     from sourcespec.ssp_plot_spectra import plot_spectra
     plot_spectra(config, spec_st, specnoise_st, plot_type='regular')
     plot_spectra(config, weight_st, plot_type='weight')
+    from sourcespec.ssp_plot_stacked_spectra import plot_stacked_spectra
+    plot_stacked_spectra(config, spec_st, sspec_output)
     from sourcespec.ssp_plot_params_stats import box_plots
     box_plots(config, sspec_output)
     if config.plot_station_map: